requirements.txt
langgraph
langchain
langchain-core
langchain-community
python-dotenv
langchain-openai
ipython
arxiv
wikipedia
langchain-tavily